requirements.txt
streamlit==1.29.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
imgaug==0.4.0
tensorflow==2.13.0
Pillow==10.2.0
scikit-image==0.20.0
mtcnn==0.1.1
tqdm==4.65.0
keras-facenet==0.3.2
deepface==0.0.79